Changed Pack to Case, for Euroleague.
Found bug in openCase where the msgSender could be an approved contract and would send the moments to the approved.
Changed to always send to the token owner.

+++ src/clancy/ERC/IClancyERC721.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.17;
 
 interface IClancyERC721 {
-    function mint() external returns (uint256);
+    function mint() external returns (uint96);
 
     // function mintTo(address to_) external returns (uint256);
 }
